feat(validator): make public and consolidate errors
- Move internal/validator to public validator/
- Consolidate internal/errors into validator/errors.go
- Move validator-specific tests to validator/validator_test.go
- Add doc.go for package documentation
