+++ validator/builtins_numeric_test.go
@@ -13,7 +13,7 @@ func TestMinMaxInvalidType(t *testing.T) {
 			Value complex128 `validate:"min=5"`
 		}
 		input := TestMin{Value: 10}
-		err := NewValidator().Struct(&input)
+		err := New().Struct(&input)
 		if err == nil {
 			t.Error("expected error for min on complex128")
 		}
@@ -23,7 +23,7 @@ func TestMinMaxInvalidType(t *testing.T) {
 			Value complex128 `validate:"max=5"`
 		}
 		input := TestMax{Value: 10}
-		err := NewValidator().Struct(&input)
+		err := New().Struct(&input)
 		if err == nil {
 			t.Error("expected error for max on complex128")
 		}
@@ -36,7 +36,7 @@ func TestValidatorErrorCases(t *testing.T) {
 			Value int `validate:"min=notanumber"`
 		}
 		input := TestMin{Value: 5}
-		err := NewValidator().Struct(&input)
+		err := New().Struct(&input)
 		if err == nil {
 			t.Error("expected error for invalid min param")
 		}
@@ -47,7 +47,7 @@ func TestValidatorErrorCases(t *testing.T) {
 			Value int `validate:"max=notanumber"`
 		}
 		input := TestMax{Value: 5}
-		err := NewValidator().Struct(&input)
+		err := New().Struct(&input)
 		if err == nil {
 			t.Error("expected error for invalid max param")
 		}
@@ -58,7 +58,7 @@ func TestValidatorErrorCases(t *testing.T) {
 			Value string `validate:"len=notanumber"`
 		}
 		input := TestLen{Value: "hello"}
-		err := NewValidator().Struct(&input)
+		err := New().Struct(&input)
 		if err == nil {
 			t.Error("expected error for invalid len param")
 		}
@@ -69,7 +69,7 @@ func TestValidatorErrorCases(t *testing.T) {
 			Value int `validate:"eq=notanumber"`
 		}
 		input := TestEq{Value: 5}
-		err := NewValidator().Struct(&input)
+		err := New().Struct(&input)
 		if err == nil {
 			t.Error("expected error for invalid eq param")
 		}
@@ -148,7 +148,7 @@ func TestUintComparisonValidators(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			err := NewValidator().Struct(&tt.input)
+			err := New().Struct(&tt.input)
 			if tt.wantErr {
 				if err == nil {
 					t.Errorf("expected error, got nil")
@@ -232,7 +232,7 @@ func TestFloatBoundaryValues(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			err := NewValidator().Struct(&tt.input)
+			err := New().Struct(&tt.input)
 			if tt.wantErr {
 				if err == nil {
 					t.Errorf("expected error, got nil")
@@ -252,7 +252,7 @@ func TestMinMaxOnUnsupportedTypes(t *testing.T) {
 			Value complex128 `validate:"min=5"`
 		}
 		input := TestMinComplex{Value: 10 + 5i}
-		err := NewValidator().Struct(&input)
+		err := New().Struct(&input)
 		if err == nil {
 			t.Error("expected error for min on complex128")
 			return
@@ -277,7 +277,7 @@ func TestMinMaxOnUnsupportedTypes(t *testing.T) {
 			Value complex128 `validate:"max=5"`
 		}
 		input := TestMaxComplex{Value: 10 + 5i}
-		err := NewValidator().Struct(&input)
+		err := New().Struct(&input)
 		if err == nil {
 			t.Error("expected error for max on complex128")
 			return
@@ -302,7 +302,7 @@ func TestMinMaxOnUnsupportedTypes(t *testing.T) {
 			Value bool `validate:"min=1"`
 		}
 		input := TestMinBool{Value: true}
-		err := NewValidator().Struct(&input)
+		err := New().Struct(&input)
 		if err == nil {
 			t.Error("expected error for min on bool")
 		}
@@ -313,7 +313,7 @@ func TestMinMaxOnUnsupportedTypes(t *testing.T) {
 			Value bool `validate:"max=1"`
 		}
 		input := TestMaxBool{Value: true}
-		err := NewValidator().Struct(&input)
+		err := New().Struct(&input)
 		if err == nil {
 			t.Error("expected error for max on bool")
 		}
@@ -391,7 +391,7 @@ func TestInvalidValidatorParams(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			input := tt.setup()
-			err := NewValidator().Struct(input)
+			err := New().Struct(input)
 			if tt.wantErr {
 				if err == nil {
 					t.Error("expected error for invalid validator param")
@@ -415,7 +415,7 @@ func TestNumericBoundaryValues(t *testing.T) {
 			MaxInt32: 2147483647,
 			MinInt32: -2147483648,
 		}
-		err := NewValidator().Struct(&input)
+		err := New().Struct(&input)
 		if err != nil {
 			t.Errorf("unexpected error: %v", err)
 		}
@@ -427,7 +427,7 @@ func TestNumericBoundaryValues(t *testing.T) {
 		}
 		// Large value should fail
 		input := TestMax{Value: 2147483647}
-		err := NewValidator().Struct(&input)
+		err := New().Struct(&input)
 		if err == nil {
 			t.Error("expected error for large value exceeding max")
 		}
@@ -438,7 +438,7 @@ func TestNumericBoundaryValues(t *testing.T) {
 			Value uint64 `validate:"gte=0,lte=18446744073709551615"`
 		}
 		input := TestUint64{Value: ^uint64(0)}
-		err := NewValidator().Struct(&input)
+		err := New().Struct(&input)
 		if err != nil {
 			t.Errorf("unexpected error: %v", err)
 		}
@@ -451,7 +451,7 @@ func TestNumericBoundaryValues(t *testing.T) {
 			Value float64 `validate:"gte=0"`
 		}
 		input := TestNaN{Value: math.NaN()}
-		err := NewValidator().Struct(&input)
+		err := New().Struct(&input)
 		// NaN < 0 returns false, so gte=0 passes (no error)
 		// Documenting current behavior - NaN is not explicitly rejected
 		if err != nil {
@@ -463,7 +463,7 @@ func TestNumericBoundaryValues(t *testing.T) {
 			Value float64 `validate:"lt=1000"`
 		}
 		input2 := TestInf{Value: math.Inf(1)}
-		err = NewValidator().Struct(&input2)
+		err = New().Struct(&input2)
 		if err == nil {
 			t.Error("expected error for +Inf")
 		}
@@ -473,7 +473,7 @@ func TestNumericBoundaryValues(t *testing.T) {
 			Value float64 `validate:"gt=-1000"`
 		}
 		input3 := TestNegInf{Value: math.Inf(-1)}
-		err = NewValidator().Struct(&input3)
+		err = New().Struct(&input3)
 		if err == nil {
 			t.Error("expected error for -Inf")
 		}
@@ -510,7 +510,7 @@ func TestNumericBoundaryValues(t *testing.T) {
 		}
 		for _, tt := range tests {
 			t.Run(tt.name, func(t *testing.T) {
-				err := NewValidator().Struct(&tt.input)
+				err := New().Struct(&tt.input)
 				if tt.wantErr {
 					if err == nil {
 						t.Error("expected error")
@@ -536,13 +536,13 @@ func TestMinMaxWithInvalidParams(t *testing.T) {
 
 	// Test min with invalid parameter
 	minInput := TestMinInvalid{Value: "hello"}
-	if err := NewValidator().Struct(&minInput); err == nil {
+	if err := New().Struct(&minInput); err == nil {
 		t.Error("expected error for min with invalid parameter")
 	}
 
 	// Test max with invalid parameter
 	maxInput := TestMaxInvalid{Value: "hello"}
-	if err := NewValidator().Struct(&maxInput); err == nil {
+	if err := New().Struct(&maxInput); err == nil {
 		t.Error("expected error for max with invalid parameter")
 	}
 }
@@ -570,62 +570,62 @@ func TestMinMaxWithCollections(t *testing.T) {
 
 	// Test min on slice - too few items
 	minSliceInput := TestMinSlice{Items: []int{1}}
-	if err := NewValidator().Struct(&minSliceInput); err == nil {
+	if err := New().Struct(&minSliceInput); err == nil {
 		t.Error("expected error for slice with less than min items")
 	}
 
 	// Test min on slice - enough items
 	minSliceInputOK := TestMinSlice{Items: []int{1, 2, 3}}
-	if err := NewValidator().Struct(&minSliceInputOK); err != nil {
+	if err := New().Struct(&minSliceInputOK); err != nil {
 		t.Errorf("unexpected error for valid slice: %v", err)
 	}
 
 	// Test max on slice - too many items
 	maxSliceInput := TestMaxSlice{Items: []int{1, 2, 3, 4}}
-	if err := NewValidator().Struct(&maxSliceInput); err == nil {
+	if err := New().Struct(&maxSliceInput); err == nil {
 		t.Error("expected error for slice with more than max items")
 	}
 
 	// Test max on slice - within limit
 	maxSliceInputOK := TestMaxSlice{Items: []int{1, 2}}
-	if err := NewValidator().Struct(&maxSliceInputOK); err != nil {
+	if err := New().Struct(&maxSliceInputOK); err != nil {
 		t.Errorf("unexpected error for valid slice: %v", err)
 	}
 
 	// Test min on array - too few items (array has fixed size, but we check length)
 	minArrayInput := TestMinArray{Items: [3]int{1, 0, 0}}
-	if err := NewValidator().Struct(&minArrayInput); err != nil {
+	if err := New().Struct(&minArrayInput); err != nil {
 		// Arrays always have fixed length, so min=2 on [3]int should pass (len=3)
 		t.Logf("Note: arrays have fixed length, validation result: %v", err)
 	}
 
 	// Test max on array
 	maxArrayInput := TestMaxArray{Items: [3]int{1, 2, 3}}
-	if err := NewValidator().Struct(&maxArrayInput); err == nil {
+	if err := New().Struct(&maxArrayInput); err == nil {
 		t.Error("expected error for array with more than max items")
 	}
 
 	// Test min on map - too few items
 	minMapInput := TestMinMap{Items: map[string]int{"a": 1}}
-	if err := NewValidator().Struct(&minMapInput); err == nil {
+	if err := New().Struct(&minMapInput); err == nil {
 		t.Error("expected error for map with less than min items")
 	}
 
 	// Test min on map - enough items
 	minMapInputOK := TestMinMap{Items: map[string]int{"a": 1, "b": 2}}
-	if err := NewValidator().Struct(&minMapInputOK); err != nil {
+	if err := New().Struct(&minMapInputOK); err != nil {
 		t.Errorf("unexpected error for valid map: %v", err)
 	}
 
 	// Test max on map - too many items
 	maxMapInput := TestMaxMap{Items: map[string]int{"a": 1, "b": 2, "c": 3}}
-	if err := NewValidator().Struct(&maxMapInput); err == nil {
+	if err := New().Struct(&maxMapInput); err == nil {
 		t.Error("expected error for map with more than max items")
 	}
 
 	// Test max on map - within limit
 	maxMapInputOK := TestMaxMap{Items: map[string]int{"a": 1}}
-	if err := NewValidator().Struct(&maxMapInputOK); err != nil {
+	if err := New().Struct(&maxMapInputOK); err != nil {
 		t.Errorf("unexpected error for valid map: %v", err)
 	}
 }
@@ -653,37 +653,37 @@ func TestMinMaxWithInvalidParamsOnNumbers(t *testing.T) {
 
 	// Test min on int with invalid parameter
 	minIntInput := TestMinIntInvalid{Value: 10}
-	if err := NewValidator().Struct(&minIntInput); err == nil {
+	if err := New().Struct(&minIntInput); err == nil {
 		t.Error("expected error for min on int with invalid parameter")
 	}
 
 	// Test max on int with invalid parameter
 	maxIntInput := TestMaxIntInvalid{Value: 10}
-	if err := NewValidator().Struct(&maxIntInput); err == nil {
+	if err := New().Struct(&maxIntInput); err == nil {
 		t.Error("expected error for max on int with invalid parameter")
 	}
 
 	// Test min on uint with invalid parameter
 	minUintInput := TestMinUintInvalid{Value: 10}
-	if err := NewValidator().Struct(&minUintInput); err == nil {
+	if err := New().Struct(&minUintInput); err == nil {
 		t.Error("expected error for min on uint with invalid parameter")
 	}
 
 	// Test max on uint with invalid parameter
 	maxUintInput := TestMaxUintInvalid{Value: 10}
-	if err := NewValidator().Struct(&maxUintInput); err == nil {
+	if err := New().Struct(&maxUintInput); err == nil {
 		t.Error("expected error for max on uint with invalid parameter")
 	}
 
 	// Test min on float with invalid parameter
 	minFloatInput := TestMinFloatInvalid{Value: 10.5}
-	if err := NewValidator().Struct(&minFloatInput); err == nil {
+	if err := New().Struct(&minFloatInput); err == nil {
 		t.Error("expected error for min on float with invalid parameter")
 	}
 
 	// Test max on float with invalid parameter
 	maxFloatInput := TestMaxFloatInvalid{Value: 10.5}
-	if err := NewValidator().Struct(&maxFloatInput); err == nil {
+	if err := New().Struct(&maxFloatInput); err == nil {
 		t.Error("expected error for max on float with invalid parameter")
 	}
 }
@@ -705,25 +705,25 @@ func TestMinMaxInvalidParamsOnCollections(t *testing.T) {
 
 	// Test min on slice with invalid parameter
 	minSliceInput := TestMinSliceInvalid{Items: []int{1, 2}}
-	if err := NewValidator().Struct(&minSliceInput); err == nil {
+	if err := New().Struct(&minSliceInput); err == nil {
 		t.Error("expected error for min on slice with invalid parameter")
 	}
 
 	// Test max on slice with invalid parameter
 	maxSliceInput := TestMaxSliceInvalid{Items: []int{1, 2}}
-	if err := NewValidator().Struct(&maxSliceInput); err == nil {
+	if err := New().Struct(&maxSliceInput); err == nil {
 		t.Error("expected error for max on slice with invalid parameter")
 	}
 
 	// Test min on map with invalid parameter
 	minMapInput := TestMinMapInvalid{Items: map[string]int{"a": 1}}
-	if err := NewValidator().Struct(&minMapInput); err == nil {
+	if err := New().Struct(&minMapInput); err == nil {
 		t.Error("expected error for min on map with invalid parameter")
 	}
 
 	// Test max on map with invalid parameter
 	maxMapInput := TestMaxMapInvalid{Items: map[string]int{"a": 1}}
-	if err := NewValidator().Struct(&maxMapInput); err == nil {
+	if err := New().Struct(&maxMapInput); err == nil {
 		t.Error("expected error for max on map with invalid parameter")
 	}
 }
